Fix some functions into our lock-free queue (see #2)
Currently, concurrent accesses to our queue invalid the behavior of
some of our functions like [length] or [drop] (even [iter]). With the
help of @polytypic, we use a [snapshot] function which gives to us an
accurate version of our queue that we can use safely even if some
domains enqueue/dequeue on it. We added a test about [length] too.

Co-authored-by: Vesa Karvonen <vesa.a.j.k@gmail.com>