Miou.yield is a non-operation (it does nothing other than "reschedule".)
This commit specifies when we want to request events from the system. In fact,
if we have "nothing to do" other than reschedule tasks, we should observe system
  events.

Semantically, this implies having only "awaits" (which will be rescheduled) or
perhaps "yields" (which will only reschedule).

However, this check does not correspond to our case for [dom0], where we need to
know whether we are only waiting for a signal from a domain.