Protect the dom0 from any exception when we compute the initial promise
In this situation, something can appear like an exception, this patch
protect us against that and still continue to kill and stop everything
at the end even if the process raised an exception: concurrent tasks on
the dom0 can raise an exception but, before this patch, we don't
clean-up resources nicely in such situation. Now, we are sure to catch,
clean & terminate the process.