Update tests with the new behavior of Miou (deletion of the Consumed state)
t04: it seems that the number of tick is between 0 and 1 now. I did not
look deeper into it. So let's fix it.
t13: the behavior of cancel changed! Now, a resolved promise can be set
as a cancelled promise even if we already got its result t13 showed the
previous behavior. Now, Miou.cancel tags any promises (even resolved
promises) as cancelled!