Cut-off resolved promises to be GCed.
It's not a memory leak as is but it can be interesting to give an opportunity to
delete the link between our resolved promises and GC roots. Instead of directly
continue after our `await_all` (and keep in the background the link with our
promises even if they are useless then), we delay the continuation to give an
opportunity for the GC to delete all of our promises (because, with this delay,
nobody keep them then).

Co-authored-by: Leandro Ostera <leandro@abstractmachine.dev>