Fix a data-race spotted by TSan and find an other way to check if a cancellation is done
Previously, we check that children of a child are done but it was an illegal
access if the child runs into another domain. This patch uses [prm.cleaned]
instead of [Sequence.is_empty child.children] to check if children of the given
child are properly cleaned.

It requires few changes about when we set to true [prm.cleaned] and be sure to
set this value when everything is cleaned (including children).