Fallback to the worker (which introspects the shared queue between domains) only when it's needed.
On benchmarks, the mutex (system) has a huge cost. This patch add a new atomic
into the pool to set to false when the shared queue is not empty. In this case,
all domains fallback to the worker function to consume all pending tasks
available into the shared queue. Otherwise, domains just consume their own
internal queues until they are empty.

We keep the behavior that when something appears into the shared queue, domains
(all domains) introspects this shared queue in the exclusive way (via the
mutex). The test30 shows the case of a cancellation across domains and where one
of our domains should consume this task from the shared queue and do, as soon as
possible, the cancellation.