Add a documentation about Unix.file_descr, files, sockets and Miou_unix
It's legitimate to want to homogenise Miou_unix with the Unix module, but the
former has other objectives than offering all the possible system syscalls. In
fact, Miou_unix is more concerned with the use of file_descr that can block
(such as sockets) but does not concern file manipulation. This commit adds
documentation on this subject so that the user understands the difference
between Miou_unix.file_descr and Unix.file_descr and why the latter can be used
just as much in the context of Miou as the former.