Fix the way we record a new syscall outside Miou (like what we do in Miou_unix)
Currently, Miou.suspend performs several effects (2, at least). They are
an opportunity to reschedule some tasks. It exists a case where a
cancellation operates before we recorded the suspension of a new syscall
in Miou. However, the upper layer (like Miou_unix) recorded the syscall
as an active one (even if we cancelled the operation).

The implication of that is a Miou.suspend on a file-descriptor, a
cancellation by a parent of this suspension but the registration of the
file-descriptor as a valid one to monitor into Unix.select. Finally,
Unix.select raises an exception because it manipulates a "Bad file
descriptor" (the cancellation may be close the file-descriptor).

This PR update the book and the documentation.

It introduces a new argument to Miou.suspend: ?fn. This function is
called when we know that **nothing** can interrupt Miou.suspend (even a
cancellation from something). So it's better to append/record the
file-descriptor in this function to be sure that we were not cancelled
by something. By this way, we avoid this strange case where we decided
to record a new file-descriptor even if we tried to cancel the task
which wants to read/write.